support/render: move json functions to httpjson package (#1375)
The keystore has been using the handler and the io functions in the hal
package for json rendering. However, the keystore doesn't respond with
hal content but just pure application/json. Responding application/hal+json in the header
without actually following the convention in the content will confuse keystore clients (and it's probably a bug).

In summary, this PR moves handler.go, handler_test.go, and io.go out of
the hal package to httpjson package and modifies the Render function so that it knows which
Content-Type to write.
