+++ services/horizon/internal/actions_data.go
@@ -25,11 +25,7 @@ func (action *DataShowAction) JSON() error {
 	action.Do(
 		action.loadParams,
 		action.loadRecord,
-		func() {
-			hal.Render(action.W, map[string]string{
-				"value": action.Data.Value,
-			})
-		},
+		func() { hal.Render(action.W, map[string]string{"value": action.Data.Value}) },
 	)
 	return action.Err
 }
